% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clear.labels}
\alias{clear.labels}
\title{clear.labels}
\usage{
clear.labels(x)
}
\arguments{
\item{list}{List of files from your environment you want to save}

\item{file}{Name of the RData file you want to save}
}
\value{

}
\description{
clears labels from your df   # src: https://stackoverflow.com/questions/2394902/remove-variable-labels-attached-with-foreign-hmisc-spss-import-functions
}
