% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%!in\%}
\alias{\%!in\%}
\title{!in}
\usage{
... \%!in\% NA
}
\arguments{
\item{list}{List of files from your environment you want to save}

\item{file}{Name of the RData file you want to save}
}
\value{

}
\description{
Negate the %in% function
}
