% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{my_rowmeans}
\alias{my_rowmeans}
\title{my_rowmeans}
\usage{
my_rowmeans(...)
}
\arguments{
\item{list}{List of files from your environment you want to save}

\item{file}{Name of the RData file you want to save}
}
\value{

}
\description{
take row means. Used in A-CAP project
}
