% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{rename}
\usage{
rename(from, to)
}
\arguments{
\item{from}{path where the files are coming from}

\item{to}{path where you want the files to go to}
}
\value{

}
\description{
This function will create a directory if it doesnt exist, then copy the file and finally rename it. Most functions I have found will move the file and therein delete it. Which this function avoids.
}
