% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{say_something}
\alias{say_something}
\title{say_something}
\usage{
say_something(message, voice)
}
\arguments{
\item{list}{List of files from your environment you want to save}

\item{file}{Name of the RData file you want to save}
}
\value{

}
\description{
Used to tap into Windows text-to-speech. Will only work on Windows-based machines.
}
