% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{spss2date}
\alias{spss2date}
\title{Convert SPSS numeric date values to \code{Date}}
\usage{
spss2date(x)
}
\arguments{
\item{x}{Numeric vector containing SPSS date values (seconds since
1582-10-14).}
}
\value{
A \code{Date} vector corresponding to \code{x}.
}
\description{
SPSS encodes dates as the number of seconds since 1582-10-14. This helper
converts those numeric values into R \code{Date} objects. See
\url{https://stackoverflow.com/questions/37880975/spss-date-format-when-imported-into-r}.
}
