% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{spss2date}
\alias{spss2date}
\title{spss2date}
\usage{
spss2date(x)
}
\arguments{
\item{list}{List of files from your environment you want to save}

\item{file}{Name of the RData file you want to save}
}
\value{

}
\description{
Used to convert SPSS dates.   # src: https://stackoverflow.com/questions/37880975/spss-date-format-when-imported-into-r
}
